+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [0.1.0-alpha27] - 2024-09-24
 
 - fix score: dataloader panic if image is zero size
+- improve score: preset more default arguments
 
 ## [0.1.0-alpha26] - 2024-09-24